Fixed bugs in the Makefile and several files to get the library to compile
Makefile: There were missing references to the proper library names plus a few
duplicate entries for said libraries. The reference to OmnicronDialogForm.resx
was not filled in and would crash dmcs with an exception.

OmnicronMenuToolStripMenuItem.cs: The namespace has been changed from
Libraries.Omnicron to Languages.Omnicron.

OmnicronDialogForm.resx: Removed execute permissions from the file. I had built
this file in windows a while ago and because of that it had execute permissions